+++ packages/isolate-daemon/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ricsam/isolate-daemon
 
+## 0.1.29
+
+### Patch Changes
+
+- add more logging when an isolate becomes disposed, and make reload hard reload the isolate instance
+- Updated dependencies
+  - @ricsam/isolate-protocol@0.1.22
+  - @ricsam/isolate-runtime@0.1.29
+  - @ricsam/isolate-playwright@0.1.22
+
 ## 0.1.28
 
 ### Patch Changes