+++ packages/playwright/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ricsam/isolate-playwright
 
+## 0.1.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @ricsam/isolate-protocol@0.1.22
+
 ## 0.1.21
 
 ### Patch Changes